Moves pages into own directory, minor formatting edits
