+++ src/App.js
@@ -2,18 +2,12 @@ import React, { Component } from 'react';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import getMuiTheme from 'material-ui/styles/getMuiTheme';
 
-import { Card, CardTitle, CardText } from 'material-ui/Card';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 
-import Navigation from './Navigation';
-import SightingFormContainer from './SightingForm/SightingFormContainer';
-import GuideContainer from './Guide/GuideContainer';
-import UserProfileContainer from './UserProfile/UserProfileContainer';
-import LoginFormContainer from './LoginForm/LoginFormContainer';
-import Login from '../pages/Login';
-import PageWrapper from '../components/PageWrapper';
-
-import './App.css';
+import Login from './pages/login';
+import Profile from './pages/profile';
+import Guide from './pages/guide';
+import Sighting from './pages/sighting';
 
 const muiTheme = getMuiTheme({
   palette: {
@@ -47,32 +41,5 @@ class App extends Component {
     );
   }
 }
-
-const Profile = () => (
-  <div>
-    <Navigation />
-    <PageWrapper>
-      <UserProfileContainer />
-    </PageWrapper>
-  </div>
-);
-
-const Sighting = () => (
-  <div>
-    <Navigation />
-    <PageWrapper>
-      <SightingFormContainer className="sighting-form" />
-    </PageWrapper>
-  </div>
-);
-
-const Guide = () => (
-  <div>
-    <Navigation />
-    <PageWrapper>
-      <GuideContainer />
-    </PageWrapper>
-  </div>
-);
 
 export default App;
