--- src/components/App.css
@@ -1,3 +1,0 @@
-.page {
-  padding: 20px;
-}
