+++ src/components/UserProfile/SightingList.jsx
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 
 const SightingList = props => {
   console.log(props);
