+++ src/pages/home.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import Navigation from '../components/Navigation';
+import PageWrapper from '../components/PageWrapper';
+
+const Home = () => (
+  <div>
+    <Navigation />
+    <PageWrapper>
+      <h1>Home</h1>
+    </PageWrapper>
+  </div>
+);
+
+export default Home;
